Building area puller
This python script pulls the building area from a pdf. We put this data in Json format since it is a universal file format used in many api's. 

User inputs are as follows:

(filename) - Input the file name as a file path. 

File Requirements are as follows:

space input pdf - 

Window pdf - 

all files must be in the same folder as this "building area puller" file.